[feat/refactor] user 도메인으로 묶기, 로그인/회원가입 url 변경 #1
