+++ src/main/java/com/eat/eat_server/config/SecurityConfig.java
@@ -1,9 +1,8 @@
-package com.eat.eat_server.Config;
+package com.eat.eat_server.config;
 
-import com.eat.eat_server.Authentication.Jwt.JwtAuthenticationFilter;
-import com.eat.eat_server.Authentication.Jwt.JwtTokenProvider;
+import com.eat.eat_server.user.jwt.JwtAuthenticationFilter;
+import com.eat.eat_server.user.jwt.JwtTokenProvider;
 
-import jakarta.servlet.DispatcherType;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
@@ -11,7 +10,6 @@ import org.springframework.http.HttpMethod;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
 import org.springframework.security.config.http.SessionCreationPolicy;
-import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.security.crypto.factory.PasswordEncoderFactories;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.security.web.SecurityFilterChain;
