+++ src/main/java/com/eat/eat_server/user/domain/Role.java
@@ -1,4 +1,4 @@
-package com.eat.eat_server.Entity;
+package com.eat.eat_server.user.domain;
 
 public enum Role {
     ROLE_USER, ROLE_ADMIN;
