+++ src/main/java/com/eat/eat_server/user/jwt/JwtAuthenticationFilter.java
@@ -1,10 +1,9 @@
-package com.eat.eat_server.Authentication.Jwt;
+package com.eat.eat_server.user.jwt;
 
 import jakarta.servlet.FilterChain;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
-import lombok.RequiredArgsConstructor;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.filter.OncePerRequestFilter;
