+++ src/main/java/com/eat/eat_server/user/service/CustomUserDetailsService.java
@@ -1,19 +1,11 @@
-package com.eat.eat_server.Authentication;
+package com.eat.eat_server.user.service;
 
-import com.eat.eat_server.Entity.User;
-import com.eat.eat_server.Repository.UserRepository;
-import lombok.Data;
+import com.eat.eat_server.user.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
-import org.springframework.security.core.GrantedAuthority;
-import org.springframework.security.core.authority.SimpleGrantedAuthority;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.stereotype.Service;
-
-import java.util.ArrayList;
-import java.util.Collection;
-import java.util.List;
 
 @Service
 @RequiredArgsConstructor
